% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ATACTrack}
\alias{ATACTrack}
\title{ATACTrack function}
\usage{
ATACTrack(
  object,
  CC = 1,
  group.by = "celltype",
  bigwig.file.list = list(),
  MOJITOO.reduction = "MOJITOO",
  Peak.assay = "Peaks",
  Peaks = NULL,
  gene.model = NULL,
  cols = (ggsci::pal_igv())(51),
  ylim.datatrack = c(0, 16),
  fontsize.geneAxis = 5,
  fontsize.geneRegion = 10,
  fontsize.datatrack = 8,
  show.legend = T,
  genome = "hg38"
)
}
\arguments{
\item{genome}{hg19 h38 mm9 mm10}
}
\value{
A grid.gTree, use grid.draw to plot
}
\description{
Top positive&negative peaks of a CC
}
\examples{
ATACTrack()
}
\keyword{ATACTrack}
