% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{ATACTrack.Seurat}
\alias{ATACTrack.Seurat}
\title{ATACTrack function}
\usage{
\method{ATACTrack}{Seurat}(
  object,
  CC = 1,
  group.by = "celltype",
  bigwig.file.list = list(),
  MOJITOO.reduction = "MOJITOO",
  Peak.assay = "Peaks",
  Peaks = NULL,
  gene.model = NULL,
  cols = (ggsci::pal_igv())(51),
  ylim.datatrack = c(0, 16),
  fontsize.geneAxis = 5,
  fontsize.geneRegion = 10,
  fontsize.datatrack = 8,
  show.legend = T,
  genome = "hg38"
)
}
\arguments{
\item{object}{Seurat object}

\item{CC}{CC index to use which MOJITOO CC}

\item{group.by}{data tracks group.by meta.data slot}

\item{bigwig.file.list}{named list of bigwig file paths, keys should be the same as unique(object@meta.data[, group.by])}

\item{MOJITOO.reduction}{MOJITOO reduction name}

\item{Peak.assay}{Peak assay name}

\item{Peaks}{peaks to display, will use CC top 2 positive&negative peaks if NULL}

\item{gene.model}{a valid gene model object that GeneRegionTrack accept, will not display if NULL}

\item{cols}{colors to show data tracks}

\item{show.legend}{if show legend}

\item{genome}{hg19 h38 mm9 mm10}
}
\value{
A grid.gTree, use grid.draw to plot
}
\description{
Plot ATACTracks with top peaks
}
\examples{
ATACTrack()
}
\keyword{ATACTrack}
