% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA.R
\name{CCA}
\alias{CCA}
\title{CCA function}
\usage{
CCA(X, Y, xname, yname)
}
\arguments{
\item{X}{nxp matrix}

\item{Y}{nxq matrix}
}
\description{
This function allows you to integrate multi-omics dim-reductions
}
\examples{
CCA()
}
\keyword{CCA}
