% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_nmf.R
\name{ConsensusNMF}
\alias{ConsensusNMF}
\title{ConsensusNMF function}
\usage{
ConsensusNMF(labels.list, random_seed = 1, maxiter = 200)
}
\arguments{
\item{labels.list}{clusters list}

\item{random_seed}{random seed set}

\item{maxiter}{maximum iterations}
}
\description{
NMF consensus clustering
}
\examples{
ConsensusNMF()
}
