% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{.addMojitooAssay.ArchRProject}
\alias{.addMojitooAssay.ArchRProject}
\title{.addMojitooAssay.ArchRProject function}
\usage{
\method{.addMojitooAssay}{ArchRProject}(object = NULL, reduction.name = "MOJITOO")
}
\arguments{
\item{object}{ArchRProject object}

\item{reduction.name}{reduction name}

\item{add.scale}{bool if add scale}
}
\description{
Add MOJITOO reduction to an Assay
}
\examples{
.addMojitooAssay()
}
\keyword{.addMojitooAssay.ArchRProject}
