% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{.addMojitooAssay}
\alias{.addMojitooAssay}
\title{.addMojitooAssay function}
\usage{
.addMojitooAssay(object = NULL, reduction.name = "MOJITOO", ...)
}
\arguments{
\item{object}{Seurat object}

\item{reduction.name}{reduction name}
}
\description{
Add MOJITOO reduction to an Assay
}
\examples{
.addMojitooAssay()
}
\keyword{.addMojitooAssay}
