% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{.addMojitooAssay.Seurat}
\alias{.addMojitooAssay.Seurat}
\title{.addMojitooAssay.Seurat function}
\usage{
\method{.addMojitooAssay}{Seurat}(object = NULL, reduction.name = "MOJITOO", add.scale = T)
}
\arguments{
\item{object}{Seurat object}

\item{reduction.name}{reduction name}

\item{add.scale}{bool if add scale}
}
\description{
Add MOJITOO reduction to an Assay
}
\examples{
.addMojitooAssay()
}
\keyword{.addMojitooAssay.Seurat}
