% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{.CCbyGenes.ArchRProject}
\alias{.CCbyGenes.ArchRProject}
\title{.CCbyGenes.ArchRProject function}
\usage{
\method{.CCbyGenes}{ArchRProject}(
  object,
  CCs = 1:3,
  topN = 10,
  RNA.assay = "RNA",
  MOJITOO.reduction = "MOJITOO",
  filter.mito = T,
  filter.ribo = T
)
}
\arguments{
\item{object}{ArchRProject object}

\item{CCs}{CC vector which CCs}

\item{topN}{topn genes to plot}

\item{RNA.assay}{RNA assay name}

\item{MOJITOO.reduction}{mojitoo reduction name}

\item{filter.mito}{bool if filter mitochondria of genes}

\item{filter.ribo}{bool if filter ribosome of genes}
}
\description{
Top positive&negative peaks of CCs
}
\examples{
.CCbyGenes()
}
\keyword{.CCbyGenes}
