% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{.CCbyPeaks}
\alias{.CCbyPeaks}
\title{.CCbyPeaks function}
\usage{
.CCbyPeaks(
  object,
  CCs = 1:3,
  topN = 10,
  Peak.assay = "Peaks",
  MOJITOO.reduction = "MOJITOO"
)
}
\arguments{
\item{object}{Seurat object}

\item{CCs}{CC vector which CCs}

\item{topN}{topn peaks to plot}

\item{MOJITOO.reduction}{mojitoo reduction name}

\item{Peaks.assay}{peak assays}
}
\description{
Top positive&negative peaks of CCs
}
\examples{
.CCbyPeaks()
}
\keyword{.CCbyPeaks.Seurat}
