% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/viz.R
\name{GeneCCDimPlot}
\alias{GeneCCDimPlot}
\alias{GeneCCDimPlot.Seurat}
\alias{GeneCCDimPlot.ArchRProject}
\title{GeneCCDimPlot function}
\usage{
GeneCCDimPlot(
  object,
  CCsToPlot = 1:3,
  RNA.assay = "RNA",
  umap = "MOJITOO_UMAP",
  MOJITOO.reduction = "MOJITOO",
  raster = F,
  combine = F,
  cols = c("blue", "grey", "red"),
  ...
)

\method{GeneCCDimPlot}{Seurat}(
  object,
  CCsToPlot = 1:3,
  RNA.assay = "RNA",
  umap = "MOJITOO_UMAP",
  MOJITOO.reduction = "MOJITOO",
  raster = F,
  combine = F,
  cols = c("blue", "grey", "red"),
  ...
)

\method{GeneCCDimPlot}{ArchRProject}(
  object,
  CCsToPlot = 1:3,
  RNA.assay = "RNA",
  umap = "MOJITOO_UMAP",
  MOJITOO.reduction = "MOJITOO",
  raster = F,
  combine = F,
  cols = c("blue", "grey", "red"),
  ...
)
}
\arguments{
\item{object}{ArchRProject Object}

\item{reduction.list}{reduction list}

\item{dims.list}{dims list}

\item{reduction.name}{reduction name}
}
\description{
CC expression DimPlot

CC expression DimPlot

CC expression DimPlot
}
\examples{
GeneCCDimPlot()
GeneCCDimPlot()
GeneCCDimPlot()
}
\keyword{GeneCCDimPlot}
