% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/util.R
\name{getCellCol}
\alias{getCellCol}
\alias{getCellCol.ArchRProject}
\alias{getCellCol.Seurat}
\title{getCellCol function}
\usage{
getCellCol(object, ...)

\method{getCellCol}{ArchRProject}(object, ...)

\method{getCellCol}{Seurat}(object)
}
\arguments{
\item{object}{ArchRProjectObject}
}
\value{
CellColDF

CellColDF

CellColDF
}
\description{
get cell column meta

get cell column meta
}
\examples{
getCellCol()
getCellCol()
}
