% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/util.R
\name{getDimRed}
\alias{getDimRed}
\alias{getDimRed.ArchRProject}
\alias{getDimRed.Seurat}
\title{getDimRed function}
\usage{
getDimRed(object, ...)

\method{getDimRed}{ArchRProject}(object, reduction, ...)

\method{getDimRed}{Seurat}(object, reduction, ...)
}
\arguments{
\item{object}{ArchRProjectObject}

\item{reduction}{reduction dataframe}
}
\value{
Dimension reduction

Dimension reduction

Dimension reduction
}
\description{
get dimension reduction

get dimension reduction
}
\examples{
getDimRed()
getDimRed()
}
