% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/util.R
\name{getMatrix}
\alias{getMatrix}
\alias{getMatrix.ArchRProject}
\alias{getMatrix.Seurat}
\title{getMatrix function}
\usage{
getMatrix(object, ...)

\method{getMatrix}{ArchRProject}(object, matrix_name, assay = NULL, ...)

\method{getMatrix}{Seurat}(object, assay = "RNA", slot = "counts")
}
\arguments{
\item{object}{SeuratObject}

\item{matrix_name}{matrix name}

\item{assay}{assay name}

\item{slot}{counts, data, scale.data}
}
\value{
Matrix

Matrix

Matrix
}
\description{
get Matrix

get Matrix
}
\examples{
getMatrix()
getMatrix()
}
