% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojitoo.R
\name{mojitoo.ArchRProject}
\alias{mojitoo.ArchRProject}
\title{mojitoo function}
\usage{
\method{mojitoo}{ArchRProject}(
  object = NULL,
  reduction.list = list(),
  dims.list = list(),
  reduction.name = "mojitoo",
  ...
)
}
\description{
This function allows you to integrate multi-omics dim-reductions
}
\examples{
mojitoo()
}
\keyword{mojitoo}
