% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mojitoo.R
\name{mojitoo}
\alias{mojitoo}
\alias{mojitoo.Seurat}
\title{mojitoo a bunch of dimensions}
\usage{
mojitoo(object, ...)

\method{mojitoo}{Seurat}(
  object = NULL,
  reduction.list = list(),
  dims.list = list(),
  reduction.name = "mojitoo",
  corr.pval = 0.05,
  assay.list = list(),
  ...
)
}
\value{
The object with mojitoo reduction
}
\description{
This function allows you to integrate multi-omics dim-reductions
}
\examples{
mojitoo()
}
\keyword{mojitoo}
