% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojitoo.R
\name{mojitoo_joint.ArchRProject}
\alias{mojitoo_joint.ArchRProject}
\title{This function allows you to integrate multi-omics dim-reductions}
\usage{
\method{mojitoo_joint}{ArchRProject}(
  object = NULL,
  a_reduction = "IterativeLSI",
  b_reduction = "Harmony",
  a_dims = 1:30,
  b_dims = 2:30,
  reduction.name = "mojitoo_joint",
  ...
)
}
\description{
This function allows you to integrate multi-omics dim-reductions
}
\examples{
mojitoo_joint()
}
\keyword{mojitoo_joint}
