% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{mojitoo_joint}
\alias{mojitoo_joint}
\title{mojitoo_joint two dimensions}
\usage{
mojitoo_joint(object, ...)
}
\value{
The object with mojitoo_joint_dimension reduction
}
\description{
mojitoo_joint two dimensions
}
