% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojitoo.R
\name{mojitoo_joint.Seurat}
\alias{mojitoo_joint.Seurat}
\title{mojitoo joint function}
\usage{
\method{mojitoo_joint}{Seurat}(
  object = NULL,
  a_reduction = "RNA_PCA",
  b_reduction = "lsi",
  a_dims = 1:29,
  b_dims = 2:40,
  reduction.name = "mojitoo_joint",
  ...
)
}
\description{
This function allows you to integrate multi-omics dim-reductions
}
\examples{
mojitoo_joint()
}
\keyword{mojitoo_joint}
