% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/util.R
\name{setCellCol}
\alias{setCellCol}
\alias{setCellCol.ArchRProject}
\alias{setCellCol.Seurat}
\title{setCellCol function}
\usage{
setCellCol(object, ...)

\method{setCellCol}{ArchRProject}(object, meta, ...)

\method{setCellCol}{Seurat}(object, meta)
}
\arguments{
\item{object}{Seurat_Obj}

\item{meta}{meta_data data.frame}
}
\value{
OBJ

object

OBJ
}
\description{
set cell column meta

set cell column meta
}
\examples{
setCellCol()
setCellCol()
}
