% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/util.R
\name{setDimRed}
\alias{setDimRed}
\alias{setDimRed.ArchRProject}
\alias{setDimRed.Seurat}
\title{setDimRed function}
\usage{
setDimRed(object, ...)

\method{setDimRed}{ArchRProject}(object, mtx, reduction.name, type = "embeddings", force = F, ...)

\method{setDimRed}{Seurat}(object, mtx, reduction.name, force = F, ...)
}
\arguments{
\item{object}{ArchRProjectObject}

\item{mtx}{reduction dataframe}

\item{reduction.name}{reduction name}

\item{type}{reducedDims, embeddings}
}
\value{
The object with dimension reduction

The object with dimension reduction

The object with dimension reduction
}
\description{
set dimension reduction

set dimension reduction
}
\examples{
setDimRed()
setDimRed()
}
