% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mojitoo.R
\name{mojitoo}
\alias{mojitoo}
\alias{mojitoo.Seurat}
\title{mojitoo a bunch of dimensions}
\usage{
mojitoo(object, ...)

\method{mojitoo}{Seurat}(
  object = NULL,
  reduction.list = list(),
  dims.list = list(),
  reduction.name = "mojitoo",
  is.reduction.center = F,
  is.reduction.scale = F,
  fdr.method = "BH",
  corr.pval = 0.05,
  keep_separated_reductions = F,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction.list}{reduction list}

\item{dims.list}{dims vector list}

\item{reduction.name}{reduction name}

\item{is.reduction.center}{bool if center the reduction}

\item{is.reduction.scale}{bool if scale the reduction}

\item{fdr.method}{fdr method, default BH}

\item{corr.pval}{correlation test 0.05}

\item{keep_separated_reductions}{bool if keep separated reductions}
}
\value{
The object with mojitoo reduction
}
\description{
This function allows you to integrate multi-omics dim-reductions
}
\examples{
mojitoo()
}
\keyword{mojitoo}
